% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_yaml.R
\name{attach_yaml2md}
\alias{attach_yaml2md}
\title{Attach yaml to md}
\usage{
attach_yaml2md(yaml, md, out = md)
}
\arguments{
\item{yaml}{A list or a string. If a list is passed to
\code{yaml}, it is treated as a yaml structure as
returned by \code{yaml::yaml.load()}. If a file path
is passed, \code{yaml::read_yaml()} will be used to
read the file.}

\item{md}{Path to the \code{.md} or \code{.Rmd} file.}

\item{out}{Path to the output \code{.md} file. Defaults
to the path specified in \code{md}, which means
overwriting the input file.}
}
\description{
Attach yaml to md
}
