% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gd.R
\name{download_gd}
\alias{download_gd}
\title{Download Google Drive file with URL}
\usage{
download_gd(url)
}
\arguments{
\item{url}{String. URL to where the file is save.
The URL may be in either
\code{https://drive.google.com/open?id=<file-id>} or
\code{https://drive.google.com/uc?export=download&id=<file-id>}
format.}
}
\value{
A list with \code{fpath}, the path to the download file
  and \code{isZip}, whether the downloaded file is zip file.
}
\description{
The file is saved as a temporary file on the system.
}
