% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_yaml.R
\name{extr_yml}
\alias{extr_yml}
\title{Extract specific variables in a yml file into yaml obj}
\usage{
extr_yml(
  file = "manual-test/author_info/joychiang.yml",
  keys = c("author", "mysite", "comment", "id")
)
}
\description{
Extract specific variables in a yml file into yaml obj
}
\keyword{internal}
