% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_yaml.R
\name{gsheet2post}
\alias{gsheet2post}
\title{Auto attach yaml header to md based on 1 input row of df}
\usage{
gsheet2post(df, post_dir_name = NULL,
  author_yml_dir = "content/author_info/")
}
\arguments{
\item{df}{A data frame with one row.}

\item{author_yml_dir}{String. Path where yaml of author information
are stored.}
}
\description{
Auto attach yaml header to md based on 1 input row of df
}
