% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipynb2md.R
\name{ipynb2md}
\alias{ipynb2md}
\title{Convert Jupyter Notebook to Markdown}
\usage{
ipynb2md(ipynb = "index.ipynb", docker = "jgoldfar/pandoc-nbconvert-docker")
}
\description{
Expected to work in dir where the source \code{.ipynb}
  is available.
}
