% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet_preprocess.R
\name{read_gs}
\alias{read_gs}
\title{Read upload response from Google Sheet}
\usage{
read_gs(url, clean_fun = collabin:::preprocess)
}
\description{
"https://docs.google.com/spreadsheets/d/1LeDRR3iXflUDgDD5WdOjXJfLXCDtdBpetna_R6rXtH0/edit?usp=sharing"
}
