% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_yaml.R
\name{rename_yaml}
\alias{rename_yaml}
\title{Rename yaml obj (a list) based on input lst.}
\usage{
rename_yaml(yaml, rename_lst = list(oldname = "newname"))
}
\description{
Rename yaml obj (a list) based on input lst.
}
\keyword{internal}
