% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zip.R
\name{unzip_ipynb}
\alias{unzip_ipynb}
\title{Unzip, find, and copy .ipynb to destination folder}
\usage{
unzip_ipynb(zipfile, postdir)
}
\arguments{
\item{zipfile}{String. Path to zip archive.}

\item{postdir}{String. Path to target directory.}
}
\description{
Unzip, find, and copy .ipynb to destination folder
}
