% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_yaml.R
\name{yaml4post}
\alias{yaml4post}
\title{Combine Two Sources of yaml Together}
\usage{
yaml4post(author_id, gs_yaml, author_yml_dir = "content/author_info/")
}
\arguments{
\item{author_id}{String. The id of the author as specified in
\enumerate{
  \item \code{content/author_info/id.yml}
  \item The filenames of the files stored in \code{content/author_info/}
    (except \code{id.yml}).
}
Note that the id is calculated from the response of the Google Form.
It normalizes the respondent-provided id by 1) conversion to lower case
and 2) replace spaces with \code{_}.}
}
\description{
This function is used to combine a 'static' yaml, saved
locally in \code{content/author_info/}, and a 'dynamic' yaml,
retrieved and computed from responses collected by Google Form.
}
