% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet_preprocess.R
\name{read_gs}
\alias{read_gs}
\title{Read & Preprocess Response from Google Sheet}
\usage{
read_gs(url, preprocess = collabin:::loper)
}
\arguments{
\item{url}{URL to Google Sheet. The sheet needs to be published
through web for it to work without authentication.}

\item{preprocess}{Function. A function passed to preprocess
the data frame returned by \code{googlesheets::gs_read()}.
It takes \code{df} as input and returns a modified \code{df}.}
}
\description{
Read pulished Google sheet into data frame and preprocess
the data frame according to the function passed in the
argument.
}
